Fix: adjust integrate delete payment method
