+++ src/utils/confirmActionToast.ts
@@ -3,17 +3,19 @@ import { toast } from "./toast";
 
 export const confirmActionToast = async (
   title: string,
-  callback: () => Promise<void>
+  callback: () => Promise<void>,
+  messageSuccess?: string
 ) => {
   Swal.fire({
     title: title,
     showCancelButton: true,
-    confirmButtonText: "Excluir",
+    confirmButtonText: "Confirmar",
     confirmButtonColor: "#00DEA3",
     cancelButtonColor: "#E85959",
+    cancelButtonText: "Cancelar"
   }).then(async (result) => {
     if (result.isConfirmed) {
-      toast("success", "Método de pagamento removido com sucesso!");
+      toast("success", messageSuccess);
       await callback();
     }
   });
